refactor(Laboratory): revalue the test to type byte

+++ src/main/java/com/corporate/talent/services/Laboratory.java
@@ -16,8 +16,8 @@ public class Laboratory {
             System.out.println("Message:" + e.getMessage());
         }
 
-        Employable empA = new Employable((byte)20, (short)10, 1000.0, 'M', 0.1f, true, 2, "Test", 80);
-        Employable empB = new Employable((byte)20, (short)10, 1000.0, 'M', 0.1f, true, 2, "Test", 80);
+        Employable empA = new Employable((byte)20, (short)10, 1000.0, 'M', 0.1f, true, 2, "Test", (byte) 80);
+        Employable empB = new Employable((byte)20, (short)10, 1000.0, 'M', 0.1f, true, 2, "Test", (byte) 80);
 
         System.out.println("\n¿empA == empB is:? " + (empA == empB));
     }
